added scoring.
Points are exponentially increased as your point multiplier increases.
Winner is still determined by most pairs matched.
I'm not sure what high score was supposed to be used for.
Also, what was called "score" was just a counter of wins for a
particular player, so I renamed it wins.. because points is more like a
score, than number of wins.